move reconnect tests to test_reconnect.py
